First pass at adding GitHub data to the graph
This is the first part of this codebase to create nodes and relationships
using the neo4j official module. This is much more Cypher focused than
the module I've previously been using but provides more direct control
of Neo4j. Even compared to the alpha level of the rest of the code in this
repo this script is raw.