Change the staff data model
Having the staff as a YAML hash prevents you from having two people with the same
name in the orggraph. This commit is the start of moving employee data down a level and
using a staff id, which must be unique, as the identifer.

The refactor then grew slightly and some of the Employee data attribute names
were changed to make them more understanable.