Move Jobs to their own data generator
The Job nodes and relationships were part of the original, very long,
script that parsed the data and loaded it into the graph. This commit
extracts those resources and converts them from their own YAML into
the intermediate format I've been using for the enrichers to make everything
more consistent and composible.