Move Employees to their own data generator
The Employee nodes and relationships were part of the original, very long,
script that parsed the data and loaded it into the graph. This commit
extracts those resources and converts them from their own YAML into
the intermediate format I've been using for the enrichers to make everything
more consistent and composible. This newer code also starts the move to work with Employee data
keyed to a staff id rather than an explicit name