Move from CREATE to MERGE
If two identical nodes with the same properties exist this
change will prevent two nodes being created. It's a very naive
first step to allow the same nodes to be specified in multiple places.
This is desirable so individual enrichers can create bare bones relationships
when nodes don't exist while also attaching to more comprehensive grpahs when
they do.

This does not correctly handle the addition of new or different sets of properties.